Make time derivatives double precision. Output weighted average of time derivatives in input device driver test.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@9085 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/include/mp/MpInputDeviceManager.h
@@ -278,7 +278,7 @@ public:
       /// @brief as relates to reference time.
     OsStatus getTimeDerivatives(MpInputDeviceHandle deviceId,
                                 unsigned& nDerivatives,
-                                int*& derivativeBuf) const;
+                                double*& derivativeBuf) const;
       /**<
       *  Calculates the derivative: 
       *  <tt>(t2-t1)/(reference frame period) for sequential t1,t2</tt>
